fix: circular import issue with config and logging (hopefully)
