+++ threeML/config/config.py
@@ -1,8 +1,23 @@
+import os
+from pathlib import Path
+
 from omegaconf import OmegaConf
 
-from threeML.io.package_data import get_path_of_user_config
-
 from .config_structure import Config
+
+
+def get_path_of_user_config() -> Path:
+    if os.environ.get("THREEML_CONFIG") is not None:
+        config_path: Path = Path(os.environ.get("THREEML_CONFIG"))
+
+    else:
+        config_path: Path = Path().home() / ".config" / "threeML"
+
+    if not config_path.exists():
+        config_path.mkdir(parents=True)
+
+    return config_path
+
 
 # Read the default Config
 threeML_config: Config = OmegaConf.structured(Config)
